Refactor explanation and reading components: implement Reader component, enhance reading progress tracking, and add view mode toggle functionality
